Shuffle some code around (NFC) and improve CUSPARSE handle creation.
