+++ src/blas/CUBLAS.jl
@@ -1,14 +1,15 @@
 module CUBLAS
 
-using CUDAdrv
-import CUDAdrv: CuStream_t
+import CUDAdrv: CUDAdrv, CuContext, CuStream_t
+
+using ..CuArrays
+using ..CuArrays: libcublas, active_context
 
 using LinearAlgebra
 
-using ..CuArrays: CuArray, CuVector, CuMatrix, CuVecOrMat,
-                  libcublas, active_context
-
 include("libcublas_types.jl")
+include("error.jl")
+include("libcublas.jl")
 
 const _handles = Dict{CuContext,cublasHandle_t}()
 const _handle = Ref{cublasHandle_t}(C_NULL)
@@ -27,10 +28,8 @@ function handle()
     return _handle[]
 end
 
-include("error.jl")
 include("util.jl")
-include("libcublas.jl")
-include("wrap.jl")
+include("wrappers.jl")
 include("highlevel.jl")
 
 end
