+++ src/rand/libcurand_types.jl
