+++ src/sparse/CUSPARSE.jl
@@ -1,13 +1,9 @@
 module CUSPARSE
 
-import CUDAdrv
-using CUDAdrv: CuContext, CuDevice
-using CUDAnative
+import CUDAdrv: CUDAdrv, CuContext, CuStream_t
 
 using ..CuArrays
-const cudaStream_t = Ptr{Nothing}
-
-using ..CuArrays: libcusparse, configured, _getindex
+using ..CuArrays: libcusparse, active_context
 
 using SparseArrays
 using LinearAlgebra
@@ -24,41 +20,30 @@ export CuSparseMatrixCSC, CuSparseMatrixCSR,
        CuSparseMatrix, AbstractCuSparseMatrix,
        CuSparseVector
 
-include("util.jl")
 include("libcusparse_types.jl")
 include("error.jl")
 include("libcusparse.jl")
 
-const libcusparse_handles = Dict{CuContext,cusparseHandle_t}()
-const libcusparse_handle = Ref{cusparseHandle_t}()
+const _handles = Dict{CuContext,cusparseHandle_t}()
+const _handle = Ref{cusparseHandle_t}()
 
-function __init__()
-    configured || return
-
-    # initialize the library when we switch devices
-    callback = (dev::CuDevice, ctx::CuContext) -> begin
-        libcusparse_handle[] = get!(libcusparse_handles, ctx) do
-            @debug "Initializing CUSPARSE for $dev"
-            handle = Ref{cusparseHandle_t}()
-            cusparseCreate(handle)
-            handle[]
+function handle()
+    if _handle[] == C_NULL
+        @assert isassigned(active_context) # some other call should have initialized CUDA
+        _handle[] = get!(_handles, active_context[]) do
+            context = active_context[]
+            handle = cusparseCreate()
+            atexit(()->CUDAdrv.isvalid(context) && cusparseDestroy(handle))
+            handle
         end
     end
-    push!(CUDAnative.device!_listeners, callback)
 
-    # deinitialize when exiting
-    atexit() do
-        libcusparse_handle[] = C_NULL
-
-        for (ctx, handle) in libcusparse_handles
-            if CUDAdrv.isvalid(ctx)
-                cusparseDestroy(handle)
-            end
-        end
-    end
+    return _handle[]
 end
 
-include("sparse.jl")
+include("array.jl")
+include("util.jl")
+include("wrappers.jl")
 include("highlevel.jl")
 
 end
