+++ test/sparse.jl
@@ -444,6 +444,7 @@ end
         end
     end
 end
+
 @testset "ic2" begin
     @testset for elty in [Float32,Float64,ComplexF32,ComplexF64]
         @testset "csr" begin
